Give marks to the things that are impala
